requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
cogwatch==2.1.0
discord.py==1.7.1
idna==3.1
multidict==5.1.0
mutagen==1.45.1
pycparser==2.20
PyNaCl==1.4.0
six==1.16.0
typing-extensions==3.10.0.0
watchgod==0.7
yarl==1.6.3
